Make data_providers a package; move seq_array
