+++ hebel/data_providers/__init__.py
@@ -1,0 +1,2 @@
+from .basic import DataProvider, MiniBatchDataProvider, BatchDataProvider, \
+    MultiTaskDataProvider, DummyDataProvider, MNISTDataProvider
